Makefile: separate build and install steps of ./lang (#60130)
Propagate LOCALE_DIR down to ./lang. The ./lang Makefile expects
LOCALE_DIR, but unless we export it, it does not propagated during `make
-C lang` calls by default.

Separate the build and install targets in the ./lang Makefile. The main
Makefile invokes `make -C lang` during the `all` target if we asked for
languages, which end up in the build directory (this was the previous
behavior anyway, as LOCALE_DIR did not propagate). The main Makefile
also now invokes `make -C lang install` specifically during `install`
targets, which copies the built mo files to LOCALE_DIR.