Improve grammar, consistency and flags of in-game windows and their descriptions (#59849)
* Improve grammar and consistency

Remove cases where objects with multiple sheets of glass stated they only had one. Made the single and multi-pane glass descriptions consistent.

* Corrected grammar mistakes

* Changed deconstruction yield

Made deconstruction yields identical to the physical description of the window

* Specify if open or closed and fix flags

Specifies whether or not a curtain or window is open or closed. Also fixes some inappropriate flags.

* Linting

* Additional flag fixes

Removes an inappropriate flag and adds in a couple I missed. Also adds in a description I forgot to change.

* Linting

* Additional linting

* Readded flag

Readded PERMEABLE flag to a window.